server/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
twilio
openai
